Made some changes to fix some things that popped up when I ran it on my windows PC
Note, you may need to install Node JS

+++ wedding_site/db/migrate/20160723032155_drop_models.rb
@@ -1,5 +1,5 @@
 class DropModels < ActiveRecord::Migration[5.0]
   def change
-  	drop_table :models
+  	drop_table :models if ActiveRecord::Base.connection.table_exists? :models
   end
 end
